Delete iy_dsm data stack method
It's the slowest by far. The IY register is used by the system and must
be preserved in the routines. The only disadvantage of ix_dsm is that
the IX register must be preserved by the caller when necessary.

FossilOrigin-Name: 9a1d01a937436edb868d8ef9f52801bdf173a644ef853aadfe551ccc6102419b